문법 수정 및 Id Long타입으로 변경

+++ NangPaGo-be/src/main/java/com/mars/NangPaGo/domain/recipelike/repository/RecipeLikeRepository.java
@@ -7,6 +7,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface RecipeLikeRepository extends JpaRepository<RecipeLike, Integer> {
+public interface RecipeLikeRepository extends JpaRepository<RecipeLike, Long> {
     Optional<RecipeLike> findByUserAndRecipe(User user, Recipe recipe);
 }
